# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workflows": {
      "post": {
        "operationId": "WorkflowsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "WorkflowsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkflowResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{id}": {
      "put": {
        "operationId": "WorkflowsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkflowRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      },
      "get": {
        "operationId": "WorkflowsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      },
      "delete": {
        "operationId": "WorkflowsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/runs": {
      "get": {
        "operationId": "WorkflowsController_listRuns",
        "parameters": [
          {
            "name": "workflowId",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List all workflow runs with metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "workflowId": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "RUNNING",
                              "COMPLETED",
                              "FAILED",
                              "CANCELLED",
                              "TERMINATED",
                              "CONTINUED_AS_NEW",
                              "TIMED_OUT",
                              "UNKNOWN"
                            ]
                          },
                          "startTime": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "endTime": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "temporalRunId": {
                            "type": "string"
                          },
                          "workflowVersionId": {
                            "type": "string",
                            "nullable": true
                          },
                          "workflowVersion": {
                            "type": "number",
                            "nullable": true
                          },
                          "workflowName": {
                            "type": "string"
                          },
                          "eventCount": {
                            "type": "number"
                          },
                          "nodeCount": {
                            "type": "number"
                          },
                          "duration": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{id}/commit": {
      "post": {
        "operationId": "WorkflowsController_commit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compiled workflow definition",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string",
                      "nullable": true
                    },
                    "entrypoint": {
                      "type": "object",
                      "properties": {
                        "ref": {
                          "type": "string"
                        }
                      }
                    },
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "ref": {
                            "type": "string"
                          },
                          "componentId": {
                            "type": "string"
                          },
                          "params": {
                            "type": "object",
                            "additionalProperties": true
                          },
                          "dependsOn": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          }
                        }
                      }
                    },
                    "config": {
                      "type": "object",
                      "properties": {
                        "environment": {
                          "type": "string"
                        },
                        "timeoutSeconds": {
                          "type": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/{id}/run": {
      "post": {
        "operationId": "WorkflowsController_run",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunWorkflowRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workflow execution result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string",
                      "description": "Temporal workflow identifier"
                    },
                    "workflowId": {
                      "type": "string",
                      "description": "Workflow record id"
                    },
                    "temporalRunId": {
                      "type": "string",
                      "description": "Temporal first execution run id"
                    },
                    "taskQueue": {
                      "type": "string",
                      "description": "Temporal task queue used for execution"
                    },
                    "workflowVersionId": {
                      "type": "string",
                      "description": "Workflow version identifier used for execution"
                    },
                    "workflowVersion": {
                      "type": "integer",
                      "description": "Workflow version number used for execution"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "RUNNING",
                        "COMPLETED",
                        "FAILED",
                        "CANCELLED",
                        "TERMINATED",
                        "CONTINUED_AS_NEW",
                        "TIMED_OUT",
                        "UNKNOWN"
                      ]
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
    "/workflows/runs/{runId}/status": {
      "get": {
        "operationId": "WorkflowsController_status",
        "parameters": [
          {
            "name": "runId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "temporalRunId",
            "required": false,
            "in": "query",
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current Temporal execution status"
          }
        },
        "tags": [
          "workflows"
        ]
      }
    },
...
  },
  "info": {
    "title": "ShipSec Studio API",
    "description": "ShipSec backend API specification",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "RunWorkflowRequestDto": {
        "type": "object",
        "properties": {
          "inputs": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "versionId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          }
        }
      },
...
    }
  }
}